src/Steam/requirements.txt
pyspark==3.4.0
delta-spark==2.4.0
python-dotenv
requests
pandas
